+++ src/test/java/com/planbase/pdf/lm2/pages/SinglePageTest.kt
@@ -145,10 +145,10 @@ val qbfCell = Cell(paleGreenLeft, squareSide,
 
 val qbfTable: Table.WrappedTable =
         Table(mutableListOf(squareSide, squareSide))
-                .partBuilder()
-                .rowBuilder()
+                .startPart()
+                .startRow()
                 .cell(paleBlueLeft, listOf(Text(times15, "The quick brown fox jumps over the lazy dog")))
                 .cell(paleBlueLeft, listOf(Text(times15, "Etaoin shrdlu cmfwyp")))
-                .buildRow()
-                .buildPart()
+                .endRow()
+                .endPart()
                 .wrap()